# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 商机随手报 API
  version: 1.0.0
  description: 固定问卷、区县承接、分侧处理、双节点审核与市公司成功商机库。
servers:
  - url: /api/v1
security:
  - cookieAuth: []
    roleGrant: []
paths:
...
  /opportunities/{opportunityId}:
    parameters: [{ $ref: "#/components/parameters/OpportunityId" }]
    get:
      operationId: getOpportunity
      responses:
        "200":
          {
            description: detail,
            content:
              {
                application/json:
                  {
                    schema: { $ref: "#/components/schemas/OpportunityDetail" },
                  },
              },
          }
        default: { $ref: "#/components/responses/Problem" }
...
  /opportunities/{opportunityId}/reviews/first/approve:
    post:
      operationId: approveFirstReview
      parameters:
        [
          { $ref: "#/components/parameters/OpportunityId" },
          { $ref: "#/components/parameters/IdempotencyKey" },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/VersionRequest" } },
            },
        }
      responses:
        {
          "201": { $ref: "#/components/responses/Transitioned" },
          default: { $ref: "#/components/responses/Problem" },
        }
  /opportunities/{opportunityId}/reviews/first/return:
    post:
      operationId: returnFirstReview
      parameters:
        [
          { $ref: "#/components/parameters/OpportunityId" },
          { $ref: "#/components/parameters/IdempotencyKey" },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/ReturnRequest" } },
            },
        }
      responses:
        {
          "201": { $ref: "#/components/responses/Transitioned" },
          default: { $ref: "#/components/responses/Problem" },
        }
  /opportunities/{opportunityId}/assignments/reassign:
    post:
      operationId: reassignOpportunity
      parameters:
        [
          { $ref: "#/components/parameters/OpportunityId" },
          { $ref: "#/components/parameters/IdempotencyKey" },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/ReassignRequest" } },
            },
        }
      responses:
        {
          "201": { $ref: "#/components/responses/Transitioned" },
          default: { $ref: "#/components/responses/Problem" },
        }
  /opportunities/{opportunityId}/pause:
    post:
      operationId: pauseOpportunity
      parameters:
        [
          { $ref: "#/components/parameters/OpportunityId" },
          { $ref: "#/components/parameters/IdempotencyKey" },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/PauseRequest" } },
            },
        }
      responses:
        {
          "201": { $ref: "#/components/responses/Transitioned" },
          default: { $ref: "#/components/responses/Problem" },
        }
  /opportunities/{opportunityId}/resume:
    post:
      operationId: resumeOpportunity
      parameters:
        [
          { $ref: "#/components/parameters/OpportunityId" },
          { $ref: "#/components/parameters/IdempotencyKey" },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/VersionRequest" } },
            },
        }
      responses:
        {
          "201": { $ref: "#/components/responses/Transitioned" },
          default: { $ref: "#/components/responses/Problem" },
        }
  /opportunities/{opportunityId}/results/success:
    post:
      operationId: submitSuccessResult
      description: 成功结果仅允许成功商机名称，不接受金额、产品、日期或其他扩展字段。
      parameters:
        [
          { $ref: "#/components/parameters/OpportunityId" },
          { $ref: "#/components/parameters/IdempotencyKey" },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema: { $ref: "#/components/schemas/SuccessResultRequest" },
                },
            },
        }
      responses:
        {
          "201": { $ref: "#/components/responses/Transitioned" },
          default: { $ref: "#/components/responses/Problem" },
        }
  /opportunities/{opportunityId}/results/failure:
    post:
      operationId: submitFailureResult
      parameters:
        [
          { $ref: "#/components/parameters/OpportunityId" },
          { $ref: "#/components/parameters/IdempotencyKey" },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema: { $ref: "#/components/schemas/FailureResultRequest" },
                },
            },
        }
      responses:
        {
          "201": { $ref: "#/components/responses/Transitioned" },
          default: { $ref: "#/components/responses/Problem" },
        }
  /opportunities/{opportunityId}/reviews/final/approve:
    post:
      operationId: approveFinalReview
      parameters:
        [
          { $ref: "#/components/parameters/OpportunityId" },
          { $ref: "#/components/parameters/IdempotencyKey" },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/VersionRequest" } },
            },
        }
      responses:
        {
          "201": { $ref: "#/components/responses/Transitioned" },
          default: { $ref: "#/components/responses/Problem" },
        }
  /opportunities/{opportunityId}/reviews/final/return:
    post:
      operationId: returnFinalReview
      parameters:
        [
          { $ref: "#/components/parameters/OpportunityId" },
          { $ref: "#/components/parameters/IdempotencyKey" },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/ReturnRequest" } },
            },
        }
      responses:
        {
          "201": { $ref: "#/components/responses/Transitioned" },
          default: { $ref: "#/components/responses/Problem" },
        }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: oms_session }
    roleGrant: { type: apiKey, in: header, name: X-Role-Grant-Id }
  parameters:
    OpportunityId:
      {
        in: path,
        name: opportunityId,
        required: true,
        schema: { type: string, format: uuid },
      }
    IdempotencyKey:
      {
        in: header,
        name: Idempotency-Key,
        required: true,
        schema: { type: string, minLength: 1, maxLength: 160 },
      }
...
  responses:
    Problem:
      description: problem details
      content:
        {
          application/problem+json:
            { schema: { $ref: "#/components/schemas/ProblemDetails" } },
        }
    Transitioned:
      description: state transitioned
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/TransitionResult" } },
        }
...
  schemas:
...
    CustomerType: { type: string, enum: [PERSONAL, ORGANIZATION] }
    CustomerAttitude:
      { type: string, enum: [URGENT, IMPORTANT, GENERAL, POTENTIAL] }
    OpportunityState:
      {
        type: string,
        enum:
          [
            PENDING_FIRST_REVIEW,
            RETURNED_TO_REPORTER,
            HANDLING,
            PAUSED,
            PENDING_FINAL_REVIEW,
            RETURNED_TO_HANDLER,
            CLOSED_SUCCESS,
            CLOSED_FAILURE,
          ],
      }
    District:
      type: object
      additionalProperties: false
      required: [id, code, name, enabled]
      properties:
        {
          id: { type: string, format: uuid },
          code: { type: string },
          name: { type: string },
          enabled: { type: boolean },
        }
...
    VersionRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties: { expectedVersion: { type: integer, minimum: 1 } }
    ReturnRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, reason]
      properties:
        {
          expectedVersion: { type: integer, minimum: 1 },
          reason: { type: string, minLength: 1, maxLength: 500 },
        }
    ReassignRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, handlerGrantId, reason]
      properties:
        {
          expectedVersion: { type: integer, minimum: 1 },
          handlerGrantId: { type: string, format: uuid },
          reason: { type: string, minLength: 1, maxLength: 500 },
        }
    PauseRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, nextProcessingAt]
      properties:
        {
          expectedVersion: { type: integer, minimum: 1 },
          nextProcessingAt: { type: string, format: date-time },
        }
    SuccessResultRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, successOpportunityName]
      properties:
        {
          expectedVersion: { type: integer, minimum: 1 },
          successOpportunityName:
            {
              type: string,
              minLength: 1,
              maxLength: 200,
              description: 请填写公司DICT系统项目名称,
            },
        }
    FailureResultRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, failureReason]
      properties:
        {
          expectedVersion: { type: integer, minimum: 1 },
          failureReason: { type: string, minLength: 1, maxLength: 500 },
        }
    TransitionResult:
      type: object
      additionalProperties: false
      required: [id, state, version]
      properties:
        {
          id: { type: string, format: uuid },
          state: { $ref: "#/components/schemas/OpportunityState" },
          version: { type: integer },
        }
    OpportunityDetail:
      type: object
      description: 字段以服务端记录级授权裁剪；市公司仅成功终态可获得。
      required:
        [
          id,
          serialNumber,
          state,
          version,
          reporterPhone,
          customerType,
          district,
          customerContact,
          specificNeed,
          attitude,
          consentAt,
        ]
      properties:
        id: { type: string, format: uuid }
        serialNumber: { type: string }
        state: { $ref: "#/components/schemas/OpportunityState" }
        version: { type: integer }
        reporterPhone: { type: string }
        customerType: { $ref: "#/components/schemas/CustomerType" }
        district: { $ref: "#/components/schemas/District" }
        customerContact: { type: string }
        specificNeed: { type: string }
        attitude: { $ref: "#/components/schemas/CustomerAttitude" }
        oneSentenceDescription: { type: [string, "null"] }
        successOpportunityName: { type: [string, "null"] }
        failureReason: { type: [string, "null"] }
        consentAt: { type: string, format: date-time }
...
    ProblemDetails:
      type: object
      required: [type, title, status, detail, errorCode]
      properties:
        {
          type: { type: string, format: uri },
          title: { type: string },
          status: { type: integer },
          detail: { type: string },
          instance: { type: string },
          errorCode: { type: string },
          traceId: { type: [string, "null"] },
        }